Note: Possible problem w/ reconcilation of bigdata-bsbm/build.xml.  In doubt, use the version from the trunk next time there is a merge from the trunk or when this branch is merged back to the trunk.

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@2786 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-bsbm/src/java/benchmark/bigdata/BigdataLoader.java
@@ -74,13 +74,24 @@ public class BigdataLoader {
                 }
             }
             
-            final File journalFile = new File(properties.getProperty(BigdataSail.Options.FILE));
-            if (journalFile.exists()) {
-                if(!journalFile.delete()) {
-                    throw new RuntimeException("could not delete old journal file");
+            final File journalFile;
+            {
+                final String tmp = properties
+                        .getProperty(BigdataSail.Options.FILE);
+                if (tmp == null) {
+                    throw new RuntimeException(
+                            "Required property not specified: "
+                                    + BigdataSail.Options.FILE);
                 }
+                journalFile = new File(tmp);
+                if (journalFile.exists()) {
+                    if (!journalFile.delete()) {
+                        throw new RuntimeException(
+                                "could not delete old journal file");
+                    }
+                }
+                System.out.println("Journal: " + journalFile);
             }
-            System.out.println("Journal: "+journalFile);
 
             // Create the journal (we know it does not exist per above).
             jnl = new Journal(properties);
